slim/context/requirements.txt
-i https://pypi.org/simple/
aiofiles==0.4.0
aniso8601==7.0.0
apispec==3.0.0
apistar==0.7.2
certifi==2019.9.11
chardet==3.0.4
click==7.0
docopt==0.6.2
graphene==2.1.8
graphql-core==2.2.1
graphql-relay==2.0.0
graphql-server-core==1.1.1
gunicorn==19.9.0
h11==0.8.1
httptools==0.0.13
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.3
markupsafe==1.1.1
marshmallow==3.2.1
numpy==1.17.3
opencv-python==4.1.1.26
parse==1.12.1
pillow==6.2.0
promise==2.2.1
python-multipart==0.0.5
pyyaml==5.1.2
requests-toolbelt==0.9.1
requests==2.22.0
responder==1.3.2
rfc3986==1.3.2
rx==1.6.1
six==1.12.0
starlette==0.12.10
typesystem==0.2.4
urllib3==1.25.6
uvicorn==0.8.6
uvloop==0.12.2 ; sys_platform != 'win32' and sys_platform != 'cygwin' and sys_platform != 'cli'
websockets==7.0
whitenoise==4.1.4
